drishtigyan/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
email-validator==2.3.0
google-generativeai==0.8.3
scipy==1.14.1
numpy==2.1.2
python-multipart==0.0.12
aiofiles==24.1.0
reportlab==4.2.5
